+++ x/committee/types/keys.go
@@ -44,7 +44,7 @@ func GetVoteKey(proposalID uint64, voter sdk.AccAddress) []byte {
 // Uint64ToBytes converts a uint64 into fixed length bytes for use in store keys.
 func uint64ToBytes(id uint64) []byte {
 	bz := make([]byte, 8)
-	binary.BigEndian.PutUint64(bz, uint64(id))
+	binary.BigEndian.PutUint64(bz, id)
 	return bz
 }
